requirements.txt
tensorflow==2.13.1
numpy<2
pillow
streamlit